Renamed middleware images folder to correspond with the convention in lib
