+++ core/server/web/middleware/image/index.js
