+++ core/server/web/middleware/image/normalize.js
